+++ internal/lang/eval.go
@@ -103,6 +103,7 @@ func (s *Scope) EvalSelfBlock(body hcl.Body, self cty.Value, schema *configschem
 	diags = diags.Append(refDiags)
 
 	terraformAttrs := map[string]cty.Value{}
+	tofuAttrs := map[string]cty.Value{}
 	pathAttrs := map[string]cty.Value{}
 
 	// We could always load the static values for Path and Terraform values,
@@ -125,6 +126,11 @@ func (s *Scope) EvalSelfBlock(body hcl.Body, self cty.Value, schema *configschem
 			diags = diags.Append(valDiags)
 			terraformAttrs[subj.Name] = val
 
+		case addrs.TofuAttr:
+			val, valDiags := normalizeRefValue(s.Data.GetTofuAttr(subj, ref.SourceRange))
+			diags = diags.Append(valDiags)
+			tofuAttrs[subj.Name] = val
+
 		case addrs.CountAttr, addrs.ForEachAttr:
 			// each and count have already been handled.
 
@@ -142,6 +148,7 @@ func (s *Scope) EvalSelfBlock(body hcl.Body, self cty.Value, schema *configschem
 
 	vals["path"] = cty.ObjectVal(pathAttrs)
 	vals["terraform"] = cty.ObjectVal(terraformAttrs)
+	vals["tofu"] = cty.ObjectVal(tofuAttrs)
 
 	ctx := &hcl.EvalContext{
 		Variables: vals,
@@ -287,6 +294,7 @@ func (s *Scope) evalContext(refs []*addrs.Reference, selfAddr addrs.Referenceabl
 	outputValues := map[string]cty.Value{}
 	pathAttrs := map[string]cty.Value{}
 	terraformAttrs := map[string]cty.Value{}
+	tofuAttrs := map[string]cty.Value{}
 	countAttrs := map[string]cty.Value{}
 	forEachAttrs := map[string]cty.Value{}
 	checkBlocks := map[string]cty.Value{}
@@ -400,6 +408,11 @@ func (s *Scope) evalContext(refs []*addrs.Reference, selfAddr addrs.Referenceabl
 			diags = diags.Append(valDiags)
 			terraformAttrs[subj.Name] = val
 
+		case addrs.TofuAttr:
+			val, valDiags := normalizeRefValue(s.Data.GetTofuAttr(subj, rng))
+			diags = diags.Append(valDiags)
+			tofuAttrs[subj.Name] = val
+
 		case addrs.CountAttr:
 			val, valDiags := normalizeRefValue(s.Data.GetCountAttr(subj, rng))
 			diags = diags.Append(valDiags)
@@ -442,6 +455,7 @@ func (s *Scope) evalContext(refs []*addrs.Reference, selfAddr addrs.Referenceabl
 	vals["local"] = cty.ObjectVal(localValues)
 	vals["path"] = cty.ObjectVal(pathAttrs)
 	vals["terraform"] = cty.ObjectVal(terraformAttrs)
+	vals["tofu"] = cty.ObjectVal(tofuAttrs)
 	vals["count"] = cty.ObjectVal(countAttrs)
 	vals["each"] = cty.ObjectVal(forEachAttrs)